requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.5
python-dotenv==1.0.1
langchain==0.3.8
langchain-openai==0.2.7
langchain-community==0.3.7
pypdf==4.3.1
playwright==1.51.0
rapidfuzz==3.10.1
